2025_04_16 commit

+++ Main/src/ExterneFDCrawler.java
@@ -5,6 +5,8 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
+import java.util.List;
+import java.util.Stack;
 
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -16,7 +18,9 @@ import utilities.Url;
 
 public class ExterneFDCrawler {
 	public static final String fs = System.getProperty("file.separator");
-	
+
+	private static List<String> content = new Stack<String>();;
+
 	static void downloadPseudoFile(String linkString, String dateiPfad) throws Exception {
 		File datei = new File(dateiPfad);
 		try {
@@ -28,31 +32,51 @@ public class ExterneFDCrawler {
 		writer.println(linkString);
 		writer.close();
 	}
-	
+
 	static void downloadFile(String linkString, String dateiPfad) throws Exception {
+		if (new File(dateiPfad).exists())
+			return;
 		InputStream in = new URL(linkString).openStream();
 		Files.copy(in, Paths.get(dateiPfad), StandardCopyOption.REPLACE_EXISTING);
+		in.close();
 	}
 
-	static void recursiveCrawl(String url, String pfad, String hauptPfad, PrintWriter md5sums) throws Exception {
+	private static void recursiveCrawl(String url, String pfad, String hauptPfad) throws Exception {
 		Document doc = Url.getWebsite(url);
-		Element content = doc.child(0);
-		Elements links = content.getElementsByTag("a");
+		Element elem = doc.child(0);
+		Elements links = elem.getElementsByTag("a");
 		for (Element link : links) {
 			String href = link.attr("href");
 			if (href.startsWith("?"))
 				continue;
 			String linkString = url.concat(href);
 			if (linkString.endsWith("/")) {
-				pfad = pfad.concat(href.substring(0, href.length() - 1)).concat(fs);
-				new File(pfad).mkdir();
-				recursiveCrawl(linkString, pfad, hauptPfad, md5sums);
+				String unterpfad = pfad.concat(href.substring(0, href.length() - 1)).concat(fs);
+				new File(unterpfad).mkdir();
+				recursiveCrawl(linkString, unterpfad, hauptPfad);
 			} else {
 				String dateiPfad = pfad.concat(href);
-				downloadPseudoFile(linkString, dateiPfad);
-				md5sums.println(Misc.md5sumOfFile(dateiPfad) + "  " + pfad.substring(hauptPfad.length()).concat(href));
 				System.out.println("Datei: " + linkString + " -> " + dateiPfad);
+				downloadFile(linkString, dateiPfad);
+				content.add(dateiPfad);
 			}
+		}
+	}
+
+	private static void calculateMd5sums(String hauptPfad) throws Exception {
+		File md5sums = new File(hauptPfad.concat("md5sums.txt"));
+		md5sums.createNewFile();
+		PrintWriter md5sumsWriter = new PrintWriter(md5sums, "UTF-8");
+		for (String dateiPfad : content) {
+			md5sumsWriter.println(Misc.md5sumOfFile(dateiPfad) + "  " + dateiPfad.substring(hauptPfad.length()));
+		}
+		md5sumsWriter.close();
+	}
+
+	private static void checkVorhanden(String hauptPfad) throws Exception {
+		for (String dateiPfad : content) {
+			if (!new File(dateiPfad).exists())
+				throw new Exception("Datei " + dateiPfad + " fehlt");
 		}
 	}
 
@@ -65,15 +89,15 @@ public class ExterneFDCrawler {
 		String contentPfad = hauptPfad.concat("content").concat(fs);
 		File pfadFile = new File(contentPfad);
 		pfadFile.mkdirs();
-		File md5sums = new File(hauptPfad.concat("md5sums.txt"));
-		md5sums.createNewFile();
-		PrintWriter md5sumsWriter = new PrintWriter(md5sums, "UTF-8");
-		recursiveCrawl(url, contentPfad, hauptPfad, md5sumsWriter);
-		md5sumsWriter.close();
+		recursiveCrawl(url, contentPfad, hauptPfad);
+		checkVorhanden(hauptPfad);
+		calculateMd5sums(hauptPfad);
 	}
 
 	public static void main(String[] args) throws Exception {
-		crawl("frl:6424451");
+//		crawl("frl:6424451");
+		crawl("frl:6425518");
+		System.out.println("ExterneFDCrawler Ende");
 	}
 
 }
